fix: Don’t keep history on gh-pages branch
It get’s big and is fairly pointless. It’s not a great way to keep track
of changes because it’s build output, not actual source. The source is
in the main project.

Long term we should move to not using the gh-pages branch, but deploying
the web site as a build asset, but that requires reconfiguring the git
project. This is a quick hack that solves the worst of it.

See: https://github.com/peaceiris/actions-gh-pages?tab=readme-ov-file#%EF%B8%8F-force-orphan-force_orphan for details.